+++ apps/chaingraph-frontend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @badaitech/chaingraph-frontend
 
+## 0.4.2
+
+### Patch Changes
+
+- feat: add betas feature to enable beta functionalities in Anthropic LLM call, ensure response stream is closed after execution
+- Updated dependencies
+  - @badaitech/chaingraph-nodes@0.4.2
+  - @badaitech/chaingraph-types@0.4.2
+  - @badaitech/badai-api@0.4.2
+  - @badaitech/chaingraph-trpc@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes